// apps/backend/src/generated/prisma/models/Advertisement.ts
export type AdvertisementWhereInput = {
  AND?: Prisma.AdvertisementWhereInput | Prisma.AdvertisementWhereInput[]
  OR?: Prisma.AdvertisementWhereInput[]
  NOT?: Prisma.AdvertisementWhereInput | Prisma.AdvertisementWhereInput[]
  id?: Prisma.StringFilter<"Advertisement"> | string
  title?: Prisma.StringFilter<"Advertisement"> | string
  description?: Prisma.StringNullableFilter<"Advertisement"> | string | null
  advertiserName?: Prisma.StringFilter<"Advertisement"> | string
  creativeUrl?: Prisma.StringFilter<"Advertisement"> | string
  destinationUrl?: Prisma.StringFilter<"Advertisement"> | string
  ctaLabel?: Prisma.StringNullableFilter<"Advertisement"> | string | null
  placement?: Prisma.EnumAdvertisementPlacementFilter<"Advertisement"> | $Enums.AdvertisementPlacement
  status?: Prisma.EnumAdvertisementStatusFilter<"Advertisement"> | $Enums.AdvertisementStatus
  startAt?: Prisma.DateTimeNullableFilter<"Advertisement"> | Date | string | null
  endAt?: Prisma.DateTimeNullableFilter<"Advertisement"> | Date | string | null
  impressionCount?: Prisma.IntFilter<"Advertisement"> | number
  clickCount?: Prisma.IntFilter<"Advertisement"> | number
  createdById?: Prisma.StringFilter<"Advertisement"> | string
  createdAt?: Prisma.DateTimeFilter<"Advertisement"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Advertisement"> | Date | string
  deletedAt?: Prisma.DateTimeNullableFilter<"Advertisement"> | Date | string | null
  createdBy?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
}